requirements.txt
pandas
openai
python-dotenv
torch
transformers
scikit-learn
tqdm